+++ src/main/kotlin/io/snyk/plugin/ui/jcef/OpenFileLoadHandlerGenerator.kt
@@ -10,12 +10,15 @@ import org.cef.browser.CefBrowser
 import org.cef.browser.CefFrame
 import org.cef.handler.CefLoadHandlerAdapter
 
+// "|" works also for separating in windows paths
+const val navigationSeparator = "|"
+
 class OpenFileLoadHandlerGenerator(
     private val project: Project,
     private val virtualFiles: LinkedHashMap<String, VirtualFile?>,
 ) {
     fun openFile(value: String): JBCefJSQuery.Response {
-        val values = value.replace("\n", "").split(":")
+        val values = value.replace("\n", "").split(navigationSeparator)
         val filePath = values[0]
         val startLine = values[1].toInt()
         val endLine = values[2].toInt()
@@ -23,11 +26,16 @@ class OpenFileLoadHandlerGenerator(
         val endCharacter = values[4].toInt()
 
         val virtualFile = virtualFiles[filePath] ?: return JBCefJSQuery.Response("success")
+        val document = virtualFile.getDocument()?: return JBCefJSQuery.Response("success")
 
-        val document = virtualFile.getDocument()
-        val startLineStartOffset = document?.getLineStartOffset(startLine) ?: 0
-        val startOffset = startLineStartOffset + (startCharacter)
-        val endLineStartOffset = document?.getLineStartOffset(endLine) ?: 0
+        // Ensure line numbers are within bounds
+        val maxLine = document.lineCount - 1
+        val safeStartLine = startLine.coerceIn(0, maxLine)
+        val safeEndLine = endLine.coerceIn(0, maxLine)
+
+        val startLineStartOffset = document.getLineStartOffset(safeStartLine)
+        val startOffset = startLineStartOffset + startCharacter
+        val endLineStartOffset = document.getLineStartOffset(safeEndLine)
         val endOffset = endLineStartOffset + endCharacter - 1
 
         navigateToSource(project, virtualFile, startOffset, endOffset)
@@ -54,10 +62,10 @@ class OpenFileLoadHandlerGenerator(
 
                         function navigateToIssue(e, target) {
                             e.preventDefault();
-                            window.openFileQuery(target.getAttribute("file-path") + ":" +
-                                 target.getAttribute("start-line") + ":" +
-                                 target.getAttribute("end-line") + ":" +
-                                 target.getAttribute("start-character") + ":" +
+                            window.openFileQuery(target.getAttribute("file-path") + "$navigationSeparator" +
+                                 target.getAttribute("start-line") + "$navigationSeparator" +
+                                 target.getAttribute("end-line") + "$navigationSeparator" +
+                                 target.getAttribute("start-character") + "$navigationSeparator" +
                                  target.getAttribute("end-character"));
                         }